Split credits between shared/fusion
